backend/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
joblib
firebase-admin
python-dotenv
python-multipart
